refactor : NextWalRepository 패키지 변경에 의한 변동
